Support pull_request_target workflow event
