+++ .nvmrc
@@ -1,1 +1,1 @@
-14.15.0
+12
