+++ package.json
@@ -43,7 +43,7 @@
     "@types/jest": "^26.0.20",
     "@types/mime": "^2.0.3",
     "@types/nock": "^11.1.0",
-    "@types/node": "^14.14.31",
+    "@types/node": "^12.20.6",
     "@typescript-eslint/parser": "^3.10.1",
     "eslint": "^7.21.0",
     "eslint-plugin-github": "^4.1.1",
