--- src/git-utils.ts
@@ -1,17 +1,0 @@
-import {Commit} from './git'
-
-export function getCommitMessage(commit: Commit): string {
-  let message = commit.message
-
-  if (message) {
-    return message
-  }
-
-  message = 'Compress image'
-
-  if (commit.files.length > 1) {
-    message += 's'
-  }
-
-  return message
-}
