+++ src/git/__tests__/index.test.ts
@@ -1,6 +1,6 @@
-import Git, {ContextEventName} from '../src/git'
-import nock, {Scope} from 'nock'
 import {Context} from '@actions/github/lib/context'
+import nock, {Scope} from 'nock'
+import Git from '..'
 
 describe('Git', () => {
   let target: Git
@@ -40,7 +40,7 @@ describe('Git', () => {
 
       test('should fetch files', async () => {
         const files = await target.getFiles(({
-          eventName: ContextEventName.Push,
+          eventName: 'push',
           payload: {
             commits: [{id: 'C1'}, {id: 'C2'}]
           },
@@ -65,49 +65,51 @@ describe('Git', () => {
       })
     })
 
-    describe('pull request', () => {
-      beforeEach(() => {
-        scope.get('/repos/OWNER/REPO/pulls/1/files').reply(200, [
-          {
-            id: 1,
-            status: 'added'
-          },
-          {
-            id: 2,
-            status: 'removed'
-          },
-          {
-            id: 3,
-            status: 'modified'
-          }
-        ])
+    for (const eventName of ['pull_request', 'pull_request_target']) {
+      describe(eventName, () => {
+        beforeEach(() => {
+          scope.get('/repos/OWNER/REPO/pulls/1/files').reply(200, [
+            {
+              id: 1,
+              status: 'added'
+            },
+            {
+              id: 2,
+              status: 'removed'
+            },
+            {
+              id: 3,
+              status: 'modified'
+            }
+          ])
+        })
+
+        test('should fetch files', async () => {
+          const files = await target.getFiles(({
+            eventName,
+            payload: {
+              number: 1
+            },
+            repo: {
+              owner: 'OWNER',
+              repo: 'REPO'
+            }
+          } as unknown) as Context)
+
+          expect(files).toEqual([
+            {
+              id: 1,
+              status: 'added'
+            },
+            {
+              id: 3,
+              status: 'modified'
+            }
+          ])
+
+          expect(scope.isDone()).toBe(true)
+        })
       })
-
-      test('should fetch files', async () => {
-        const files = await target.getFiles(({
-          eventName: ContextEventName.PullRequest,
-          payload: {
-            number: 1
-          },
-          repo: {
-            owner: 'OWNER',
-            repo: 'REPO'
-          }
-        } as unknown) as Context)
-
-        expect(files).toEqual([
-          {
-            id: 1,
-            status: 'added'
-          },
-          {
-            id: 3,
-            status: 'modified'
-          }
-        ])
-
-        expect(scope.isDone()).toBe(true)
-      })
-    })
+    }
   })
 })
